requirements.txt
# Hyperliquid SDK
hyperliquid-python-sdk>=0.21.0

# Auth
eth-account>=0.11.0

# Config
python-dotenv>=1.0.0

# CLI
click>=8.1.0
rich>=13.7.0

# Logging
structlog>=24.1.0

# Scanner
requests>=2.31.0
tabulate>=0.9.0

# Backtesting
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0
matplotlib>=3.8.0
tqdm>=4.66.0
